Configurato servizio IQuerable

+++ src/MyCourse/Models/Entities/Course.cs
@@ -8,37 +8,35 @@ namespace MyCourse.Models.Entities
     {
         public Course(string title, string author)
         {
-            if (string.IsNullOrWhiteSpace(title))
+		    if (string.IsNullOrWhiteSpace(title))
             {
                 throw new ArgumentException("The course must have a title");
             }
-            if (string.IsNullOrWhiteSpace(author))
+			if (string.IsNullOrWhiteSpace(author))
             {
                 throw new ArgumentException("The course must have an author");
-            }
-            
+            }   
 
             Title = title;
             Author = author;
             Lessons = new HashSet<Lesson>();
         }
 
-        public long Id { get; private set; }
-        public string Title { get;private set; }
-        public string Description { get;private set; }
-        public string ImagePath { get;private set; }
-        public string Author { get;private set; }
-        public string Email { get;private set; }
-        public double Rating { get;private set; }
-        public Money FullPrice{ get;private set; }
-        public Money CurrentPrice { get;private set; }
-        
+        public int Id { get; private set; }
+        public string Title { get; private set; }
+        public string Description { get; private set; }
+        public string ImagePath { get; private set; }
+        public string Author { get; private set; }
+        public string Email { get; private set; }
+        public double Rating { get; private set; }
+        public Money FullPrice { get; private set; }
+        public Money CurrentPrice { get; private set; }
 
         public void ChangeTitle(string newTitle)
         {
             if (string.IsNullOrWhiteSpace(newTitle))
             {
-              throw new ArgumentException("The course must have a title");  
+                throw new ArgumentException("The course must have a title");
             }
             Title = newTitle;
         }
@@ -60,8 +58,7 @@ namespace MyCourse.Models.Entities
             FullPrice = newFullPrice;
             CurrentPrice = newDiscountPrice;
         }
-        
 
-        public virtual ICollection<Lesson> Lessons { get;private set; }
+        public virtual ICollection<Lesson> Lessons { get; private set; }
     }
 }
